CI: move topology related tests to new module
The tests ensuring the correctness of the CPU
and cache topology were moved to a new file
named "test_topology" for readability reasons.
